requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
django-mssql-backend==2.8.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Markdown==3.6
mssql-django==1.5
PyJWT==2.8.0
pyodbc==5.1.0
pytz==2024.1
sqlparse==0.5.0
tzdata==2024.1
whitenoise==6.6.0